Revamped Player Movement
Revamped how Player Movement works, as well as re-adding the Knight Idle animation, and revamping how the Map objects are structured (You'll see what I mean when you open it up).

PS. From now on please add in a "Platform" tag to every platform we add into the game, it will allow the player to jump on said platform!